add Game_Object and Object_Manager class
they are probaply renamend in future, but the main idee behind them is, that an object contains all the logict to create a verteci with texture and the object manager has them in a list, tracks there positions. and other things. Probaply also the function to compile shader and load shaders and use them. But how this is down needs to be found out in detail