feat: feature project cleanup (#5)
* BUG: multible cubes stacked have no collison

when multible cubes are spawned in row, collison triggers but does not launch car back

* dirty quick fix for multible collision

* add simple rectancle map out of multible cubes

* initial cleanup commit

This is the first commit for the Project cleanup, with this the Game was moved into the Game Project, for Guys with Engines, they need to figure out on how to run this project now, entrypoint is Program.cs, the next stepps will be to move OpenGL related stuff to the Engine and write hellper function so the ppl using in the funture don't have to think too much about OpenGL / OpenTK

* seperated first Engine and Game components

Moves things like the shader and texture class to the Engine Namespace and created a GameApp namespace which will contain the game the Engine namespace should contin things to make working with OpenTK / OpenGL more easy in the future, like abstakting creating VBOs VAOs and binding them, at the end there should be wrapper classes were you can just load a texture on to the screen and position it without having tho think about OpenTK / OpenGL stuff

* remove old tutorial content

* add Game_Object and Object_Manager class

they are probaply renamend in future, but the main idee behind them is, that an object contains all the logict to create a verteci with texture and the object manager has them in a list, tracks there positions. and other things. Probaply also the function to compile shader and load shaders and use them. But how this is down needs to be found out in detail

* add initial idee on GameObjects and GameObjectManagement
The idee is that the GameObjects will contain all the informtaion on how
to draw them, meaning things like VBO, VAO, EBO, Shader, Texture,
position and so on. With this the ObjectManager will then be able to
setup everything to draw the GameObject and Draw it. It also will
contain a List of all GameObject objects in the list can be added,
deleted, and updated. So in therory, the game logic like colission and
other ings can use the Objects in the list to do there stuff

* update Engine Project structure
Update Engine Project structure to be somewhat more organiced, will
probpaly change.

TODO: Add static Classes like Shapes.cs to contains usefull shapes

TODO: Add static Classes like WindowInput to make Inputs global
available

TODO: Add static GameState Class

* update GameObject and ObjectManger
The Object manager is now able to add del and return a GameObject form
the list.

The structure of the GameObject is not final and will change

* minor fixes.
TODO: think of a way to make the Shape class more flexible, and how to
pass the shape type flexible to a GameObject

* add resource manager, fix minior logic errors

* finish resource manager add a bit more logik to the Engine
Next up will be a simple game to test the Engine

* basic setup works now, can draw car and container as background
ToDo: figure out how to load a map, and find the alpha values, nad match
thme to the postion of the car in the world. to then do collision and
other stuff

* got basic fotn working. problems with shaders
and other shit i don't get at the time of dooing this shit and i need to
be done in about 24h and there is still no map builder fuck fuck fuck

* delete UI_classes as do not work
Need to figure out a way to recreate the texture renderer from learn
OpenGL

* Add new textures and restructure resources folder

* add ttf loader still needs to be tested.

ToDo: Add actuall text render function, and the corosbonding shaders
ToDo: Go through code and remove not needed comments and probaply a
little cleanup on basic structure

* working text on screen. with font

ToDo: Implement Map builder

* add layout for map builder

* add working collision dedection on the Map

---------

Co-authored-by: Leonhard Schaefler <leonhard.schaefler@gmail.com>
Co-authored-by: Nico <87312762+Nicolai-Geisler@users.noreply.github.com>